refactor(agent): extract activity/memory/sanitization mixins from run_agent.py (shard s3)
