+++ run_agent.py
@@ -222,6 +222,13 @@ from agent.tool_dispatch_helpers import (
 )
 from utils import atomic_json_write, base_url_host_matches, base_url_hostname, env_float, is_truthy_value, model_forces_max_completion_tokens
 
+from plugins.agent.mixins.activity_mixin import ActivityMixin
+from plugins.agent.mixins.memory_lifecycle_mixin import MemoryLifecycleMixin
+from plugins.agent.mixins.message_sanitization_mixin import MessageSanitizationMixin
+from plugins.agent.mixins.todo_hydration_mixin import TodoHydrationMixin
+from plugins.agent.mixins.turn_footers_mixin import TurnFootersMixin
+from plugins.agent.mixins.usage_telemetry_mixin import UsageTelemetryMixin
+
 
 # Internal flags that mark a message as ephemeral empty-response/prefill
 # recovery scaffolding: the synthetic assistant "(empty)" turn and user nudge
@@ -409,7 +416,14 @@ class _StreamErrorEvent(Exception):
         }
 
 
-class AIAgent:
+class AIAgent(
+    ActivityMixin,
+    MemoryLifecycleMixin,
+    MessageSanitizationMixin,
+    TodoHydrationMixin,
+    TurnFootersMixin,
+    UsageTelemetryMixin,
+):
     """
     AI Agent with tool calling capabilities.
 
@@ -3382,761 +3396,35 @@ class AIAgent:
             self._pending_steer = None
         return text
 
-    def _record_file_mutation_result(
-        self,
-        tool_name: str,
-        args: Dict[str, Any],
-        result: Any,
-        is_error: bool,
-    ) -> None:
-        """Record a ``write_file`` / ``patch`` outcome for the turn-end verifier.
 
-        On failure, store ``{path: {error_preview, tool}}`` entries.  On
-        success, remove any prior failure entries for the same paths (the
-        model recovered within the turn).  Silently no-ops if the per-turn
-        state dict hasn't been initialised yet (e.g. a tool dispatched
-        outside ``run_conversation``).
-        """
-        if tool_name not in _FILE_MUTATING_TOOLS:
-            return
-        state = getattr(self, "_turn_failed_file_mutations", None)
-        if state is None:
-            return
-        targets = _extract_file_mutation_targets(tool_name, args)
-        if not targets:
-            return
-        landed = file_mutation_result_landed(tool_name, result)
-        if landed:
-            changed = getattr(self, "_turn_file_mutation_paths", None)
-            if changed is not None:
-                changed.update(_extract_landed_file_mutation_paths(tool_name, args, result))
-        if is_error and not landed:
-            preview = _extract_error_preview(result)
-            for path in targets:
-                # Keep the FIRST error we saw for a given path unless we
-                # later see success.  A repeated failure with a different
-                # message shouldn't silently overwrite the original.
-                if path not in state:
-                    state[path] = {
-                        "tool": tool_name,
-                        "error_preview": preview,
-                    }
-        else:
-            for path in targets:
-                state.pop(path, None)
-
-    def _file_mutation_verifier_enabled(self) -> bool:
-        """Check whether the per-turn file-mutation verifier footer is on.
-
-        Config path: ``display.file_mutation_verifier`` (bool, default True).
-        ``HERMES_FILE_MUTATION_VERIFIER`` env var overrides config.  Exposed
-        as a method so tests can patch a single seam without reaching into
-        the private ``_turn_failed_file_mutations`` state dict.
-        """
-        try:
-            import os as _os
-            env = _os.environ.get("HERMES_FILE_MUTATION_VERIFIER")
-            if env is not None:
-                return env.strip().lower() not in {"0", "false", "no", "off"}
-            # Read from the persisted config.yaml so gateway and CLI share
-            # the same setting.  Import lazily to avoid a startup-time cycle.
-            try:
-                from hermes_cli.config import load_config as _load_config
-                _cfg = _load_config() or {}
-            except Exception:
-                _cfg = {}
-            _display = _cfg.get("display") if isinstance(_cfg, dict) else None
-            if isinstance(_display, dict) and "file_mutation_verifier" in _display:
-                return bool(_display.get("file_mutation_verifier"))
-        except Exception:
-            pass
-        return True  # safe default: verifier on
-
-    # Bare absolute / home / Windows-drive file paths in a footer line.
-    # Anchors mirror the gateway's ``extract_local_files`` bare-path
-    # detector so that anything the gateway WOULD auto-attach is wrapped
-    # in inline-code backticks here first (the extractor skips paths inside
-    # `code` spans).  Defense-in-depth: even if a future error message
-    # echoes a credential path (config.yaml, .env, auth.json) into the
-    # user-facing footer, it can never be matched as a deliverable bare
-    # path and silently uploaded to a messaging channel (#35584).
     _FOOTER_PATH_RE = re.compile(
         r"(?<![/:\w.`])(?:~/|/|[A-Za-z]:[/\\])(?:[\w.\-]+[/\\])*[\w.\-]+\.[\w]+",
     )
 
-    @classmethod
-    def _neutralize_footer_paths(cls, text: str) -> str:
-        """Wrap bare file paths in backticks so they aren't auto-delivered.
 
-        The gateway's ``extract_local_files`` scans response text for bare
-        absolute/home paths ending in a deliverable extension and uploads
-        any that exist on disk as native attachments — but it explicitly
-        skips paths inside inline-code (`` `...` ``) spans.  Backticking
-        every path the footer renders defeats that auto-detection while
-        keeping the path fully human-readable.  Paths already wrapped in a
-        backtick (the negative lookbehind excludes a preceding `` ` ``) are
-        left untouched so we never double-wrap.
-        """
-        if not text:
-            return text
-        return cls._FOOTER_PATH_RE.sub(lambda m: f"`{m.group(0)}`", text)
 
-    @classmethod
-    def _format_file_mutation_failure_footer(cls, failed: Dict[str, Dict[str, Any]]) -> str:
-        """Render the per-turn failed-mutation dict as a user-facing footer.
 
-        Displays up to 10 paths with their first error preview, then a
-        count of any additional failures.  Returns an empty string when
-        the dict is empty so callers can concatenate unconditionally.
-
-        Every file path that reaches the user-facing text — both the bullet
-        path and any path echoed inside the tool's error preview — is
-        backtick-wrapped via ``_neutralize_footer_paths`` so the gateway's
-        bare-path media extractor can never auto-attach a protected file
-        (e.g. ``~/.hermes/config.yaml``) to a messaging channel (#35584).
-        """
-        if not failed:
-            return ""
-        lines = [
-            "⚠️ File-mutation verifier: "
-            f"{len(failed)} file(s) were NOT modified this turn despite any "
-            "wording above that may suggest otherwise. Run `git status` or "
-            "`read_file` to confirm."
-        ]
-        shown = 0
-        for path, info in failed.items():
-            if shown >= 10:
-                break
-            preview = (info.get("error_preview") or "").strip()
-            tool = info.get("tool") or "patch"
-            if preview:
-                lines.append(f"  • `{path}` — [{tool}] {preview}")
-            else:
-                lines.append(f"  • `{path}` — [{tool}] failed")
-            shown += 1
-        remaining = len(failed) - shown
-        if remaining > 0:
-            lines.append(f"  • … and {remaining} more")
-        # Neutralize any path the preview text echoed (the bullet path is
-        # already backticked above; the lookbehind keeps it from being
-        # double-wrapped).
-        return cls._neutralize_footer_paths("\n".join(lines))
-
-    def _turn_completion_explainer_enabled(self) -> bool:
-        """Check whether the end-of-turn completion explainer footer is on.
-
-        Config path: ``display.turn_completion_explainer`` (bool, default
-        True).  ``HERMES_TURN_COMPLETION_EXPLAINER`` env var overrides
-        config.  Exposed as a method so tests can patch a single seam,
-        mirroring ``_file_mutation_verifier_enabled``.
-        """
-        try:
-            import os as _os
-            env = _os.environ.get("HERMES_TURN_COMPLETION_EXPLAINER")
-            if env is not None:
-                return env.strip().lower() not in {"0", "false", "no", "off"}
-            # Read from the persisted config.yaml so gateway and CLI share
-            # the same setting.  Import lazily to avoid a startup-time cycle.
-            try:
-                from hermes_cli.config import load_config as _load_config
-                _cfg = _load_config() or {}
-            except Exception:
-                _cfg = {}
-            _display = _cfg.get("display") if isinstance(_cfg, dict) else None
-            if isinstance(_display, dict) and "turn_completion_explainer" in _display:
-                return bool(_display.get("turn_completion_explainer"))
-        except Exception:
-            pass
-        return True  # safe default: explainer on
-
-    @staticmethod
-    def _format_turn_completion_explanation(turn_exit_reason: str) -> str:
-        """Render a user-facing explanation for an abnormal turn ending.
-
-        Maps the internal ``turn_exit_reason`` to a short, actionable
-        message so a turn that produced no usable assistant reply (empty
-        content after retries, a partial/truncated stream, a still-pending
-        tool result, or an iteration/budget limit) is never silent from
-        the UI's perspective — the symptom users report in #34452.
-
-        Returns an empty string for reasons that are NOT abnormal (e.g.
-        a normal ``text_response(...)`` exit), so callers can concatenate
-        or substitute unconditionally without warning on healthy turns
-        like a terse ``Done.``.
-        """
-        if not turn_exit_reason:
-            return ""
-        reason = str(turn_exit_reason)
-
-        # Normal completion — stay quiet.  ``text_response(...)`` is the
-        # healthy terminal; anything that produced a real reply is fine.
-        if reason.startswith("text_response"):
-            return ""
-
-        prefix = "⚠️ No reply: "
-        if reason == "empty_response_exhausted":
-            return (
-                prefix
-                + "the model returned empty content after retries and any "
-                "fallback providers. Try `continue`, switch model/provider, "
-                "or inspect the tool output above."
-            )
-        if reason == "all_retries_exhausted_no_response":
-            return (
-                prefix
-                + "all API retries were exhausted before a response was "
-                "produced (provider errors / rate limits). Try `continue` "
-                "or switch provider."
-            )
-        if reason == "partial_stream_recovery":
-            return (
-                prefix
-                + "streaming stopped early and only a partial response was "
-                "recovered. Send `continue` to resume from where it stopped."
-            )
-        if reason == "fallback_prior_turn_content":
-            return (
-                prefix
-                + "no new content was produced this turn; showing recovered "
-                "prior context. Send `continue` to retry."
-            )
-        if reason == "interrupted_during_api_call":
-            return (
-                prefix
-                + "the request was interrupted mid-call before a reply was "
-                "received. Send `continue` to retry."
-            )
-        if reason == "budget_exhausted":
-            return (
-                prefix
-                + "the per-turn iteration/cost budget was exhausted before a "
-                "final answer. Send `continue` to keep going."
-            )
-        if reason == "ollama_runtime_context_too_small":
-            return (
-                prefix
-                + "the local model's context window was too small to finish. "
-                "Increase the context size or use a larger model."
-            )
-        if reason.startswith("max_iterations_reached"):
-            return (
-                prefix
-                + "the maximum tool-iteration limit was reached before a "
-                "final answer. Send `continue` to keep going, or raise "
-                "`max_iterations`."
-            )
-        if reason.startswith("error_near_max_iterations"):
-            return (
-                prefix
-                + "an error occurred near the iteration limit before a final "
-                "answer. Check the tool output above, then send `continue`."
-            )
-        if reason == "pending_tool_result":
-            return (
-                prefix
-                + "the turn stopped while a tool result was still pending and "
-                "the model produced no follow-up text. Send `continue` to "
-                "let it summarize."
-            )
-        if reason == "session_persistence_failed":
-            return (
-                prefix
-                + "the turn was stopped because session storage could not be "
-                "written (the transcript would have been lost on restart). "
-                "This is often a full disk — free some space (or fix state.db "
-                "permissions), then send your message again."
-            )
-        # Unknown/diagnostic-only reasons (e.g. "unknown", guardrail_halt
-        # which already surfaces its own message) — don't second-guess.
-        return ""
 
     def _apply_pending_steer_to_tool_results(self, messages: list, num_tool_msgs: int) -> None:
         """Forwarder — see ``agent.agent_runtime_helpers.apply_pending_steer_to_tool_results``."""
         from agent.agent_runtime_helpers import apply_pending_steer_to_tool_results
         return apply_pending_steer_to_tool_results(self, messages, num_tool_msgs)
 
-    def _touch_activity(
-        self,
-        desc: str,
-        *,
-        provenance: Optional[ActivityProvenance] = None,
-        force_persist: bool = False,
-    ) -> None:
-        """Update the last-activity timestamp and description (thread-safe).
 
-        Also bridges to the kanban board's heartbeat fields when this
-        process is a dispatcher-spawned worker (HERMES_KANBAN_TASK set),
-        so the dispatcher watchdog doesn't reclaim an actively-running
-        worker as stale (#31752). Bridge is rate-limited (60s) and
-        best-effort — it never raises into the agent loop.
 
-        Separately, rate-limits a durable SessionDB activity projection
-        (``last_activity_at`` + bounded description/provenance) so
-        CLI/Gateway consumers share one observation source (#72016 / #72039).
 
-        ``provenance`` defaults to ``unknown`` (the ordinary agent activity
-        clock). Named values are for special writers (e.g. compression);
-        ordinary call sites should leave the default.
 
-        ``force_persist`` bypasses the 60s SessionDB rate limit so a
-        terminal stamp (e.g. compression completed) is not dropped.
-        """
-        from agent.session_activity import (
-            bound_activity_description,
-            normalize_activity_provenance,
-            reset_session_activity_persist_window,
-        )
 
-        self._last_activity_ts = time.time()
-        self._last_activity_desc = bound_activity_description(desc)
-        self._last_activity_provenance = normalize_activity_provenance(provenance)
-        if os.environ.get("HERMES_KANBAN_TASK"):
-            try:
-                from tools.kanban_tools import (
-                    heartbeat_current_worker_from_env,
-                    inject_new_comments_from_env,
-                )
-                heartbeat_current_worker_from_env()
-                # Fold any new operator notes into the running turn (OUT-OF-BAND
-                # steer) so the user can talk to a live task without a restart.
-                inject_new_comments_from_env(self)
-            except Exception:
-                # Never let the bridge break the agent loop.  The function
-                # already swallows exceptions internally; this outer guard
-                # covers import-time failures (kanban_tools unavailable,
-                # etc.) on niche deployment surfaces.
-                pass
-        if force_persist:
-            reset_session_activity_persist_window(self)
-        self._persist_session_activity_if_due()
 
-    def _persist_session_activity_if_due(self) -> None:
-        """Best-effort durable activity heartbeat for SessionDB consumers.
 
-        Cadence is pinned by SESSION_ACTIVITY_HEARTBEAT_MIN_INTERVAL_SECONDS
-        (>=30s per session, config-independent — see agent/session_activity.py).
-        The write rides the standard SessionDB ``_execute_write`` patience
-        path via ``touch_session_activity``. Fail-open: a failed heartbeat
-        write must NEVER raise into the agent loop (swallow + debug-log).
-        """
-        session_id = getattr(self, "session_id", None)
-        session_db = getattr(self, "_session_db", None)
-        if not session_id or session_db is None:
-            return
-        touch = getattr(session_db, "touch_session_activity", None)
-        if not callable(touch):
-            return
-        from agent.session_activity import (
-            SESSION_ACTIVITY_HEARTBEAT_MIN_INTERVAL_SECONDS,
-            normalize_activity_provenance,
-        )
 
-        now_mono = time.monotonic()
-        last_mono = getattr(self, "_session_activity_last_persist_mono", 0.0)
-        if (now_mono - last_mono) < SESSION_ACTIVITY_HEARTBEAT_MIN_INTERVAL_SECONDS:
-            return
-        self._session_activity_last_persist_mono = now_mono
-        try:
-            touch(
-                session_id,
-                getattr(self, "_last_activity_ts", None),
-                description=getattr(self, "_last_activity_desc", None),
-                provenance=normalize_activity_provenance(
-                    getattr(self, "_last_activity_provenance", None)
-                ),
-            )
-        except Exception:
-            # Never let durable heartbeat I/O break the agent loop. The
-            # heartbeat is an observation-only projection; the next due
-            # window retries naturally.
-            logger.debug(
-                "session activity heartbeat write failed (ignored)",
-                exc_info=True,
-            )
 
-    def _reset_activity_labels_after_turn(self) -> None:
-        """Drop mid-turn activity labels once the turn is no longer running.
 
-        Keeps ``_last_activity_ts`` so idle/watchdog clocks stay continuous
-        across interrupt-recursive turns (#15654) and between turns. Clears
-        description + provenance so idle cached agents / SessionDB listings
-        do not keep advertising the last mid-turn stamp (e.g. compression
-        or tool execution) after the turn ended (#72039).
-        """
-        from agent.session_activity import ActivityProvenance
 
-        self._last_activity_desc = ""
-        self._last_activity_provenance = ActivityProvenance.UNKNOWN
-        session_id = getattr(self, "session_id", None)
-        session_db = getattr(self, "_session_db", None)
-        if not session_id or session_db is None:
-            return
-        clear = getattr(session_db, "clear_session_activity_labels", None)
-        if not callable(clear):
-            return
-        try:
-            clear(session_id)
-        except Exception:
-            # Never let durable cleanup I/O break turn teardown.
-            pass
 
-    def _capture_rate_limits(self, http_response: Any) -> None:
-        """Parse x-ratelimit-* headers from an HTTP response and cache the state.
 
-        Called after each streaming API call.  The httpx Response object is
-        available on the OpenAI SDK Stream via ``stream.response``.
-        """
-        if http_response is None:
-            return
-        headers = getattr(http_response, "headers", None)
-        if not headers:
-            return
-        try:
-            from agent.rate_limit_tracker import parse_rate_limit_headers
-            state = parse_rate_limit_headers(headers, provider=self.provider)
-            if state is not None:
-                self._rate_limit_state = state
-        except Exception:
-            pass  # Never let header parsing break the agent loop
 
-    def get_rate_limit_state(self):
-        """Return the last captured RateLimitState, or None."""
-        return self._rate_limit_state
 
-    def _capture_anthropic_response_headers(self, http_response: Any) -> None:
-        """Capture out-of-band state from Anthropic Messages response headers.
-
-        The Anthropic SDK's aggregated ``Message`` drops HTTP headers. Portal
-        (and other providers) put rate-limit and credits state there — the same
-        families the OpenAI-wire streaming path captures via
-        ``stream.response``. Fail-open: each capture swallows its own errors.
-        """
-        self._capture_rate_limits(http_response)
-        self._capture_credits(http_response)
-
-    def _capture_credits(self, http_response: Any) -> None:
-        """Parse x-nous-credits-* headers, cache CreditsState, fire threshold notices.
-
-        Fail-open throughout — header issues never break the agent loop. The PARSE is
-        swallowed (any error → treated as a miss → keep last-known). The notice
-        EVALUATION/EMIT is a SEPARATE block that WARNS on failure (R1-M2): a bug in the
-        depletion-notice path must not vanish silently under the parse swallow.
-        """
-        # Dev test fixture (HERMES_DEV_CREDITS_FIXTURE): inject a chosen notice state
-        # each turn for repeatable testing, bypassing real headers. Throwaway scaffolding.
-        try:
-            from agent.credits_tracker import dev_fixture_credits_state
-            _fixture = dev_fixture_credits_state()
-        except Exception:
-            _fixture = None
-        if _fixture is not None:
-            self._credits_state = _fixture
-            if self._credits_session_start_micros is None:
-                self._credits_session_start_micros = _fixture.remaining_micros
-            _latch = getattr(self, "_credits_latch", None)
-            if isinstance(_latch, dict):
-                # Only seen_below_90 — never seen_grant_unspent (priming it would
-                # fire grant_spent on a fixture's first observation, the exact
-                # every-session nag the gate exists to prevent).
-                _latch["seen_below_90"] = True  # let warn90 fire without a real crossing
-            _used = _fixture.used_fraction
-            logger.info(
-                "credits ▸ [FIXTURE] remaining=%d (%s) · paid=%s · denom=%s · used=%s "
-                "(real headers bypassed — `echo clear` / unset HERMES_DEV_CREDITS_FIXTURE to restore)",
-                _fixture.remaining_micros,
-                _fixture.remaining_usd or "?",
-                _fixture.paid_access,
-                _fixture.denominator_kind,
-                ("%.0f%%" % (_used * 100)) if _used is not None else "n/a",
-            )
-            self._emit_credits_notices()
-            return
-        if http_response is None:
-            return
-        headers = getattr(http_response, "headers", None)
-        if not headers:
-            return
-        _dev = is_truthy_value(os.environ.get("HERMES_DEV_CREDITS"))
-
-        # ── Parse (fail-open → miss; never overwrite good state with None) ──
-        try:
-            from agent.credits_tracker import parse_credits_headers
-            state = parse_credits_headers(headers, provider=self.provider)
-        except Exception:
-            return  # parse error → treat as a miss, keep last-known
-        if state is None:
-            if _dev:
-                logger.info(
-                    "credits ▸ response had no valid x-nous-credits-* headers "
-                    "(miss — producer off / non-Nous path / >TTL stale)"
-                )
-            return
-
-        # retain-last-known: only overwrite on a fresh valid parse
-        self._credits_state = state
-        # Latch session-start remaining the first time we ever see a header
-        if self._credits_session_start_micros is None:
-            self._credits_session_start_micros = state.remaining_micros
-        if _dev:
-            # HERMES_DEV_CREDITS: stream each capture to agent.log — watch live with
-            # `hermes logs -f` (grep 'credits ▸'). Dev-only; silent for normal users.
-            spent = self.get_credits_spent_micros()
-            used = state.used_fraction
-            logger.info(
-                "credits ▸ remaining=%d (%s) · paid=%s · denom=%s · used=%s "
-                "· Δspent=%s · age=%s%s",
-                state.remaining_micros,
-                state.remaining_usd or "?",
-                state.paid_access,
-                state.denominator_kind,
-                ("%.0f%%" % (used * 100)) if used is not None else "n/a",
-                ("%.1f¢" % (spent / 10000)) if spent is not None else "n/a",
-                ("%.0fs" % state.age_seconds) if state.age_seconds != float("inf") else "n/a",
-                (" · disabled=%s" % state.disabled_reason) if state.disabled_reason else "",
-            )
-
-        # Threshold notices — shared with the cold-start seed (see _emit_credits_notices).
-        self._emit_credits_notices()
-
-    def _emit_credits_notices(self) -> None:
-        """Run the threshold policy on the current credits state and emit notices.
-
-        Shared by the warm path (_capture_credits) and the L3 cold-start seed, so a
-        session that opens already depleted warns immediately — not only after the first
-        inference header. Runs only when a notice consumer is bound (messaging binds none
-        → state still cached for /usage, no policy). WARNS on failure rather than
-        swallowing (R1-M2): a depletion-path bug must not vanish silently. Emits clears
-        FIRST, then shows (so depleted lands last in a latest-wins slot).
-        """
-        if getattr(self, "notice_callback", None) is None and getattr(self, "notice_clear_callback", None) is None:
-            return
-        if not self._credits_notices_enabled():
-            return
-        state = getattr(self, "_credits_state", None)
-        if state is None:
-            return
-        try:
-            from agent.credits_tracker import evaluate_credits_notices, is_free_tier_model, new_credits_latch
-            latch = getattr(self, "_credits_latch", None)
-            if latch is None:
-                latch = self._credits_latch = new_credits_latch()
-            # Free-model gate: a depleted account on a free model can still
-            # inference, so the depleted error banner is suppressed. Local-data
-            # only (":free" suffix + pricing-cache peek) — never a network call.
-            model_is_free = is_free_tier_model(
-                getattr(self, "model", "") or "",
-                getattr(self, "base_url", "") or "",
-            )
-            to_show, to_clear = evaluate_credits_notices(state, latch, model_is_free=model_is_free)
-            for key in to_clear:        # clears FIRST …
-                self._emit_notice_clear(key)
-            for notice in to_show:      # … then shows (depleted lands last in a latest-wins slot)
-                self._emit_notice(notice)
-        except Exception:
-            logger.warning("credits notice evaluation/emit failed", exc_info=True)
-
-    def _credits_notices_enabled(self) -> bool:
-        """Whether credits notices are enabled (config display.credits_notices).
-
-        Read once per agent and cached — the policy runs after every API
-        response, and the setting governs UI noise, not correctness, so a
-        config flip applying on the next session is fine.  Fail-open True
-        (preserve current behaviour) on any config error.
-        """
-        cached = getattr(self, "_credits_notices_enabled_cache", None)
-        if cached is not None:
-            return cached
-        enabled = True
-        try:
-            from hermes_cli.config import load_config as _load_config
-            _cfg = _load_config() or {}
-            _display = _cfg.get("display") if isinstance(_cfg, dict) else None
-            if isinstance(_display, dict) and "credits_notices" in _display:
-                enabled = bool(_display.get("credits_notices"))
-        except Exception:
-            enabled = True
-        self._credits_notices_enabled_cache = enabled
-        return enabled
-
-    def get_credits_state(self):
-        """Return the last captured CreditsState, or None."""
-        return self._credits_state
-
-    def get_credits_spent_micros(self):
-        """Session-cumulative micros spent = first_seen_remaining - current_remaining. None if no data."""
-        if self._credits_session_start_micros is None or self._credits_state is None:
-            return None
-        return self._credits_session_start_micros - self._credits_state.remaining_micros
-
-    def _check_openrouter_cache_status(self, http_response: Any) -> None:
-        """Read X-OpenRouter-Cache-Status from response headers and log it.
-
-        Increments ``_or_cache_hits`` on HIT so callers can report savings.
-        """
-        if http_response is None:
-            return
-        headers = getattr(http_response, "headers", None)
-        if not headers:
-            return
-        try:
-            status = headers.get("x-openrouter-cache-status")
-            if not status:
-                return
-            if status.upper() == "HIT":
-                self._or_cache_hits += 1
-                logger.info("OpenRouter response cache HIT (total: %d)", self._or_cache_hits)
-            else:
-                logger.debug("OpenRouter response cache %s", status.upper())
-        except Exception:
-            pass  # Never let header parsing break the agent loop
-
-    def get_activity_summary(self) -> dict:
-        """Return a snapshot of the agent's current activity for diagnostics.
-
-        Exposes the shared activity observation contract
-        (``last_activity_at`` / ``last_activity_description`` /
-        ``last_activity_provenance``) plus short aliases
-        (``last_activity_ts`` / ``last_activity_desc`` / …) for existing
-        gateway and delegate readers.
-        """
-        from agent.session_activity import (
-            ActivityProvenance,
-            build_activity_snapshot,
-        )
-
-        provenance = getattr(self, "_last_activity_provenance", None)
-        if provenance is None:
-            provenance = ActivityProvenance.UNKNOWN
-        return build_activity_snapshot(
-            last_activity_at=getattr(self, "_last_activity_ts", None),
-            last_activity_description=getattr(self, "_last_activity_desc", None) or "",
-            last_activity_provenance=provenance,
-            extra={
-            "current_tool": self._current_tool,
-            "api_call_count": self._api_call_count,
-            "max_iterations": self.max_iterations,
-            "budget_used": self.iteration_budget.used,
-            "budget_max": self.iteration_budget.max_total,
-            },
-        )
-
-    def shutdown_memory_provider(self, messages: list = None) -> None:
-        """Shut down the memory provider and context engine — call at actual session boundaries.
-
-        This calls on_session_end() then shutdown_all() on the memory
-        manager, and on_session_end() on the context engine.
-        NOT called per-turn — only at CLI exit, /reset, gateway
-        session expiry, etc.
-        """
-        if self._memory_manager:
-            try:
-                self._memory_manager.on_session_end(messages or [])
-            except Exception as e:
-                logger.warning("Memory provider on_session_end failed during shutdown: %s", e, exc_info=True)
-            try:
-                self._memory_manager.shutdown_all()
-            except Exception:
-                pass
-        # Notify context engine of session end (flush DAG, close DBs, etc.)
-        if hasattr(self, "context_compressor") and self.context_compressor:
-            try:
-                self.context_compressor.on_session_end(
-                    self.session_id or "",
-                    messages or [],
-                )
-            except Exception:
-                pass
-
-    def commit_memory_session(self, messages: list = None) -> None:
-        """Trigger end-of-session extraction without tearing providers down.
-        Called when session_id rotates (e.g. /new, context compression);
-        providers keep their state and continue running under the old
-        session_id — they just flush pending extraction now."""
-        if self._memory_manager:
-            try:
-                self._memory_manager.on_session_end(messages or [])
-            except Exception:
-                pass
-        # Notify context engine of session end too — same lifecycle moment as
-        # the memory manager's on_session_end. Without this, engines that
-        # accumulate per-session state (DAGs, summaries) leak that state from
-        # the rotated-out session into whatever comes next under the same
-        # compressor instance. Mirrors the call in shutdown_memory_provider().
-        # See issue #22394.
-        if hasattr(self, "context_compressor") and self.context_compressor:
-            try:
-                self.context_compressor.on_session_end(
-                    self.session_id or "",
-                    messages or [],
-                )
-            except Exception:
-                pass
-
-    def _sync_external_memory_for_turn(
-        self,
-        *,
-        original_user_message: Any,
-        final_response: Any,
-        interrupted: bool,
-        messages: list | None = None,
-    ) -> None:
-        """Mirror a completed turn into external memory providers.
-
-        Called at the end of ``run_conversation`` with the cleaned user
-        message (``original_user_message``) and the finalised assistant
-        response.  The external memory backend gets both ``sync_all`` (to
-        persist the exchange) and ``queue_prefetch_all`` (to start
-        warming context for the next turn) in one shot.
-
-        Uses ``original_user_message`` rather than ``user_message``
-        because the latter may carry injected skill content that bloats
-        or breaks provider queries.
-
-        Interrupted turns are skipped entirely (#15218).  A partial
-        assistant output, an aborted tool chain, or a mid-stream reset
-        is not durable conversational truth — mirroring it into an
-        external memory backend pollutes future recall with state the
-        user never saw completed.  The prefetch is gated on the same
-        flag: the user's next message is almost certainly a retry of
-        the same intent, and a prefetch keyed on the interrupted turn
-        would fire against stale context.
-
-        Normal completed turns still sync as before.  The whole body is
-        wrapped in ``try/except Exception`` because external memory
-        providers are strictly best-effort — a misconfigured or offline
-        backend must not block the user from seeing their response.
-        """
-        if interrupted:
-            return
-        if not (self._memory_manager and final_response and original_user_message):
-            return
-        # Multimodal turns carry content as a list of typed parts; providers
-        # expect plain strings, so flatten to text first (newline-joined for
-        # memory, vs the default space-join used for log/trajectory previews).
-        user_text = _summarize_user_message_for_log(original_user_message, sep="\n")
-        response_text = _summarize_user_message_for_log(final_response, sep="\n")
-        if not (user_text and response_text):
-            return
-        try:
-            sync_kwargs = {"session_id": self.session_id or ""}
-            if messages is not None:
-                sync_kwargs["messages"] = messages
-            self._memory_manager.sync_all(
-                user_text,
-                response_text,
-                **sync_kwargs,
-            )
-            # Sibling of the build_turn_context() prefetch gate: warming the
-            # next turn's recall with a trivial prompt ("hi", "thanks") keys
-            # provider searches on zero-signal text — skip it. The sync above
-            # still runs so the turn itself is persisted.
-            if not is_trivial_prompt(user_text):
-                self._memory_manager.queue_prefetch_all(
-                    user_text,
-                    session_id=self.session_id or "",
-                )
-        except Exception:
-            pass
 
     def release_clients(self) -> None:
         """Release LLM client resources WITHOUT tearing down session tool state.
@@ -4325,107 +3613,8 @@ class AIAgent:
         except Exception:
             pass
 
-    def _hydrate_todo_store(self, history: List[Dict[str, Any]]) -> None:
-        """
-        Recover todo state from conversation history.
-        
-        The gateway creates a fresh AIAgent per message, so the in-memory
-        TodoStore is empty. We scan the history for the most recent todo
-        tool response and replay it to reconstruct the state.
 
-        Hydration is restricted to tool results that are paired with an
-        earlier assistant ``todo`` tool call. The gateway/API server accepts
-        caller-supplied ``conversation_history``, so a forged bare
-        ``role: tool`` message carrying a ``todos`` array must not be able to
-        seed the store without a matching canonical tool call
-        (GHSA-5g4g-6jrg-mw3g).
-        """
-        from tools.todo_tool import MAX_TODO_RESULT_CHARS
 
-        # Walk history backwards to find the most recent todo tool response
-        last_todo_response = None
-        for idx in range(len(history) - 1, -1, -1):
-            msg = history[idx]
-            if msg.get("role") != "tool":
-                continue
-            content = msg.get("content", "")
-            if not isinstance(content, str):
-                continue
-            # Only accept tool results paired with a prior assistant todo call.
-            if not self._tool_response_matches_todo_call(history, idx):
-                continue
-            if len(content) > MAX_TODO_RESULT_CHARS:
-                logger.warning(
-                    "Skipping oversized todo tool response during hydration: "
-                    "session=%s chars=%d",
-                    self.session_id or "none",
-                    len(content),
-                )
-                continue
-            # Quick check: todo responses contain "todos" key
-            if '"todos"' not in content:
-                continue
-            try:
-                data = json.loads(content)
-                if "todos" in data and isinstance(data["todos"], list):
-                    last_todo_response = data["todos"]
-                    break
-            except (json.JSONDecodeError, TypeError):
-                continue
-
-        if last_todo_response:
-            # Replay the items into the store (replace mode)
-            self._todo_store.write(last_todo_response, merge=False)
-            if not self.quiet_mode:
-                self._vprint(f"{self.log_prefix}📋 Restored {len(last_todo_response)} todo item(s) from history")
-        _set_interrupt(False)
-
-    @classmethod
-    def _tool_response_matches_todo_call(
-        cls,
-        history: List[Dict[str, Any]],
-        tool_index: int,
-    ) -> bool:
-        """Return True when a tool result belongs to a prior assistant todo call.
-
-        Scans backwards from the tool result to the nearest assistant message
-        and confirms it issued a ``todo`` tool call whose id matches this
-        result's ``tool_call_id``. A ``user``/``system`` boundary (or a missing
-        id) means the result is unpaired and must not hydrate the store.
-        """
-        if tool_index < 0 or tool_index >= len(history):
-            return False
-        tool_msg = history[tool_index]
-        tool_call_id = tool_msg.get("tool_call_id")
-        if not tool_call_id:
-            return False
-
-        for prior_idx in range(tool_index - 1, -1, -1):
-            prior = history[prior_idx]
-            role = prior.get("role")
-            if role == "assistant":
-                return cls._assistant_has_todo_tool_call(prior, tool_call_id)
-            if role in {"user", "system"}:
-                return False
-        return False
-
-    @classmethod
-    def _assistant_has_todo_tool_call(
-        cls,
-        assistant_msg: Dict[str, Any],
-        tool_call_id: str,
-    ) -> bool:
-        """True when the assistant message issued a ``todo`` call with this id."""
-        tool_calls = assistant_msg.get("tool_calls")
-        if not isinstance(tool_calls, list):
-            return False
-
-        for tool_call in tool_calls:
-            if cls._get_tool_call_id_static(tool_call) != tool_call_id:
-                continue
-            if cls._get_tool_call_name_static(tool_call) == "todo":
-                return True
-        return False
 
     @property
     def is_interrupted(self) -> bool:
@@ -4441,224 +3630,22 @@ class AIAgent:
 
 
 
-    def _build_system_prompt_parts(self, system_message: str = None) -> Dict[str, str]:
-        """Forwarder — see ``agent.system_prompt.build_system_prompt_parts``."""
-        from agent.system_prompt import build_system_prompt_parts
-        return build_system_prompt_parts(self, system_message=system_message)
 
-    def _build_system_prompt(self, system_message: str = None) -> str:
-        """Forwarder — see ``agent.system_prompt.build_system_prompt``."""
-        from agent.system_prompt import build_system_prompt
-        return build_system_prompt(self, system_message=system_message)
 
-    @staticmethod
-    def _get_tool_call_id_static(tc) -> str:
-        """Extract call ID from a tool_call entry (dict or object).
 
-        Forwarder — policy owner is
-        ``agent.message_sanitization.coalesce_tool_call_id`` (audit F4).
-        """
-        return _sanitize_coalesce_tool_call_id(tc)
-
-    @staticmethod
-    def _get_tool_call_name_static(tc) -> str:
-        """Extract function name from a tool_call entry (dict or object).
-
-        Gemini's OpenAI-compatibility endpoint requires every `role: tool`
-        message to carry the matching function name. OpenAI/Anthropic/ollama
-        tolerate its absence, so the field is best-effort: callers fall back
-        to "" and the message still works elsewhere.
-        """
-        if isinstance(tc, dict):
-            fn = tc.get("function")
-            if isinstance(fn, dict):
-                return fn.get("name", "") or ""
-            return ""
-        fn = getattr(tc, "function", None)
-        return getattr(fn, "name", "") or ""
 
     _VALID_API_ROLES = frozenset({"system", "user", "assistant", "tool", "function", "developer"})
 
-    @staticmethod
-    def _sanitize_api_messages(messages: List[Dict[str, Any]]) -> List[Dict[str, Any]]:
-        """Forwarder — see ``agent.agent_runtime_helpers.sanitize_api_messages``."""
-        from agent.agent_runtime_helpers import sanitize_api_messages
-        return sanitize_api_messages(messages)
 
-    @staticmethod
-    def _is_thinking_only_assistant(
-        msg: Dict[str, Any],
-        *,
-        drop_codex_reasoning_items: bool = True,
-    ) -> bool:
-        """Return True if ``msg`` is an assistant turn whose only payload is reasoning.
 
-        "Thinking-only" means the model emitted reasoning (``reasoning`` or
-        ``reasoning_content``) but no visible text and no tool_calls. When sent
-        back to providers that convert reasoning into thinking blocks (native
-        Anthropic, OpenRouter Anthropic, third-party Anthropic-compatible
-        gateways), the resulting message has only thinking blocks — which
-        Anthropic rejects with HTTP 400 "The final block in an assistant
-        message cannot be `thinking`."
 
-        Symmetric with Claude Code's ``filterOrphanedThinkingOnlyMessages``
-        (src/utils/messages.ts). We drop the whole turn from the API copy
-        rather than fabricating stub text — the message log (UI transcript)
-        keeps the reasoning block; only the wire copy is cleaned.
-        """
-        if not isinstance(msg, dict) or msg.get("role") != "assistant":
-            return False
-        if msg.get("tool_calls"):
-            return False
-        # Does it have any actual output?
-        content = msg.get("content")
-        if isinstance(content, str):
-            if content.strip():
-                return False
-        elif isinstance(content, list):
-            for block in content:
-                if not isinstance(block, dict):
-                    if block:  # non-empty non-dict string etc.
-                        return False
-                    continue
-                btype = block.get("type")
-                if btype in {"thinking", "redacted_thinking"}:
-                    continue
-                if btype == "text":
-                    text = block.get("text", "")
-                    if isinstance(text, str) and text.strip():
-                        return False
-                    continue
-                # tool_use, image, document, etc. — real payload
-                return False
-        elif content is not None and content != "":
-            return False
-        # Content is empty-ish. Is there reasoning to make it thinking-only?
-        reasoning = msg.get("reasoning_content") or msg.get("reasoning")
-        if isinstance(reasoning, str) and reasoning.strip():
-            return True
-        # reasoning_details list form
-        rd = msg.get("reasoning_details")
-        if isinstance(rd, list) and rd:
-            return True
-        # Codex Responses stores encrypted reasoning state under a separate
-        # assistant-message key. Treat only real reasoning items as
-        # thinking-only; empty/junk lists should fall through to the generic
-        # empty-turn handling instead of being dropped here.
-        codex_items = msg.get("codex_reasoning_items")
-        if drop_codex_reasoning_items and isinstance(codex_items, list):
-            return any(
-                isinstance(item, dict) and item.get("type") == "reasoning"
-                for item in codex_items
-            )
-        return False
 
-    @staticmethod
-    def _drop_thinking_only_and_merge_users(
-        messages: List[Dict[str, Any]],
-        *,
-        drop_codex_reasoning_items: bool = True,
-    ) -> List[Dict[str, Any]]:
-        """Forwarder — see ``agent.agent_runtime_helpers.drop_thinking_only_and_merge_users``."""
-        from agent.agent_runtime_helpers import drop_thinking_only_and_merge_users
-        return drop_thinking_only_and_merge_users(
-            messages,
-            drop_codex_reasoning_items=drop_codex_reasoning_items,
-        )
 
-    @staticmethod
-    def _cap_delegate_task_calls(tool_calls: list) -> list:
-        """Truncate excess delegate_task calls to max_concurrent_children.
 
-        The delegate_tool caps the task list inside a single call, but the
-        model can emit multiple separate delegate_task tool_calls in one
-        turn.  This truncates the excess, preserving all non-delegate calls.
 
-        Returns the original list if no truncation was needed.
-        """
-        from tools.delegate_tool import _get_max_concurrent_children
-        max_children = _get_max_concurrent_children()
-        delegate_count = sum(1 for tc in tool_calls if tc.function.name == "delegate_task")
-        if delegate_count <= max_children:
-            return tool_calls
-        kept_delegates = 0
-        truncated = []
-        for tc in tool_calls:
-            if tc.function.name == "delegate_task":
-                if kept_delegates < max_children:
-                    truncated.append(tc)
-                    kept_delegates += 1
-            else:
-                truncated.append(tc)
-        logger.warning(
-            "Truncated %d excess delegate_task call(s) to enforce "
-            "max_concurrent_children=%d limit",
-            delegate_count - max_children, max_children,
-        )
-        return truncated
 
-    @staticmethod
-    def _deduplicate_tool_calls(tool_calls: list) -> list:
-        """Remove duplicate (tool_name, arguments) pairs within a single turn.
 
-        Only the first occurrence of each unique pair is kept.
-        Returns the original list if no duplicates were found.
-        """
-        seen: set = set()
-        unique: list = []
-        for tc in tool_calls:
-            key = (tc.function.name, tc.function.arguments)
-            if key not in seen:
-                seen.add(key)
-                unique.append(tc)
-            else:
-                logger.warning("Removed duplicate tool call: %s", tc.function.name)
-        return unique if len(unique) < len(tool_calls) else tool_calls
 
-    @staticmethod
-    def _uniquify_tool_call_ids(tool_calls: list) -> list:
-        """Ensure every tool call in a single assistant turn has a distinct id.
-
-        Forwarder — policy owner is
-        ``agent.message_sanitization.uniquify_tool_call_ids`` (audit F4).
-        First occurrence keeps its id; later collisions get a deterministic
-        ``<id>_d<n>`` suffix (never uuid4 — prompt-cache prefix stability).
-        Mutates entries in place and returns the same list.
-        """
-        return _sanitize_uniquify_tool_call_ids(tool_calls)
-
-    def _repair_tool_call(self, tool_name: str) -> str | None:
-        """Forwarder — see ``agent.agent_runtime_helpers.repair_tool_call``."""
-        from agent.agent_runtime_helpers import repair_tool_call
-        return repair_tool_call(self, tool_name)
-
-    def _invalidate_system_prompt(self):
-        """Forwarder — see ``agent.system_prompt.invalidate_system_prompt``."""
-        from agent.system_prompt import invalidate_system_prompt
-        invalidate_system_prompt(self)
-
-    @staticmethod
-    def _deterministic_call_id(fn_name: str, arguments: str, index: int = 0) -> str:
-        """Generate a deterministic call_id from tool call content.
-
-        Used as a fallback when the API doesn't provide a call_id.
-        Deterministic IDs prevent cache invalidation — random UUIDs would
-        make every API call's prefix unique, breaking OpenAI's prompt cache.
-        """
-        return _codex_deterministic_call_id(fn_name, arguments, index)
-
-    @staticmethod
-    def _split_responses_tool_id(raw_id: Any) -> tuple[Optional[str], Optional[str]]:
-        """Split a stored tool id into (call_id, response_item_id)."""
-        return _codex_split_responses_tool_id(raw_id)
-
-    def _derive_responses_function_call_id(
-        self,
-        call_id: str,
-        response_item_id: Optional[str] = None,
-    ) -> str:
-        """Build a valid Responses `function_call.id` (must start with `fc_`)."""
-        return _codex_derive_responses_function_call_id(call_id, response_item_id)
 
     def _thread_identity(self) -> str:
         thread = threading.current_thread()
